Moved commonSearchUtil.js and commonUtil.js to util/common. Created progressManageUtil.js(unused)
