+++ util/common/commonSearchUtil.js
