+++ util/common/commonUtil.js
