+++ util/common/progressManageUtil.js
@@ -1,0 +1,3 @@
+
+exports.test = "test";
+
